requirements.txt
Django==2.1.7
django-cors-middleware==1.3.1
djangorestframework==3.9.2
drf-writable-nested==0.5.1
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
